Eliminate build warnings: GTK4 API migrations, version targeting, MSVC fixes
Migrate deprecated GTK/GLib APIs to their modern replacements:
- GtkFileChooserDialog → GtkFileDialog (async callbacks)
- GtkDialog/GtkMessageDialog → GtkAlertDialog and GtkWindow
- GtkComboBoxText → GtkDropDown/GtkStringList
- GtkTreeStore → GListStore (chanview data layer rewrite)
- GCompletion → custom HcCompletion (fkeys.c tab completion)
- GTimeVal → gint64 via g_get_real_time() (DCC, util, fishlim)
- gtk_widget_show/hide → deletion or gtk_widget_set_visible()
- gtk_widget_get_allocation → gtk_widget_get_width/height
- gtk_css_provider_load_from_data → load_from_string
- Per-widget CSS label coloring → PangoAttrList attributes
- Remove inline/dummy variable from gtk-helpers.h

Additional fixes:
- Set version targets: min GLib 2.80/GTK 4.14, max GLib 2.82/GTK 4.16
- Fix config.h redefinition warnings (include ordering)
- Add HC_IGNORE_RESULT macro for unchecked I/O return values
- MSVC C4305: add f suffix to float literals (palette.c, xtext.c)
- MSVC C4244: explicit casts for SOCKET/gsize/ptrdiff_t → int
- MSVC C4013: add missing plugin-tray.h include
- Remove unused variables (menu.c, userlistgui.c, oauth.c, servlistgui.c)

+++ src/common/modes.c
@@ -206,7 +206,7 @@ nick_access (server * serv, char *nick, int *modechars)
 		nick++;
 	}
 
-	*modechars = nick - orig;
+	*modechars = (int)(nick - orig);
 
 	return access;
 }
@@ -258,10 +258,10 @@ record_chan_mode (session *sess, char sign, char mode, char *arg)
 	/* find out if the mode currently exists */
 	arguments_start = g_strstr_len(current->str	, -1, " ");
 	if (arguments_start) {
-		modes_length = arguments_start - current->str;
+		modes_length = (gint)(arguments_start - current->str);
 	}
 	else {
-		modes_length = current->len;
+		modes_length = (gint) current->len;
 		/* set this to the end of the modes */
 		arguments_start = current->str + current->len;
 	}
@@ -304,7 +304,7 @@ record_chan_mode (session *sess, char sign, char mode, char *arg)
 			if (i != argument_num)
 				current_char++;
 		}
-		argument_offset = current_char - current->str;
+		argument_offset = (gint)(current_char - current->str);
 
 		/* how long the existing argument is for this key
 		 * important for malloc and strncpy */
